Refactor user onboarding by moving greeting messages and friendship features to a new OnboardChat service
