+++ src/main/java/com/vibez/engine/Service/Implementation/UserImplement.java
@@ -10,10 +10,8 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
-import com.vibez.engine.Model.Message;
 import com.vibez.engine.Model.User;
 import com.vibez.engine.Repository.UserRepo;
-import com.vibez.engine.Service.FriendshipService;
 import com.vibez.engine.Service.JwtService;
 import com.vibez.engine.Service.UserService;
 
@@ -25,9 +23,6 @@ public class UserImplement implements UserService {
 
     @Autowired
     private JwtService jwtService;
-
-    @Autowired
-    private FriendshipService FriendshipService;
 
     @Autowired
     AuthenticationManager authenticationManager;
@@ -50,24 +45,7 @@ public class UserImplement implements UserService {
         newUser.setGroupIds(groups);
         newUser.setPassword(hashedPassword);
         userRepo.save(newUser);
-        String vibezShip = FriendshipService.sendFriendRequest(newUser.getUserId(), "679b99e09b263c53d878d30f");
-        FriendshipService.acceptFriendRequest(vibezShip);
-        greetingMessage(newUser.getUserId(), "🎉 Hey there! Welcome to VIBEZ! 🎊 We're so happy to have you here! 💜");
-        greetingMessage(newUser.getUserId(), "Now that you've joined, let me give you a quick tour! 🚀");
-        greetingMessage(newUser.getUserId(), "🔎 You can connect with others by searching for their username or email. Send them a request, and once they accept, you're all set to chat! 💬✨");
-        greetingMessage(newUser.getUserId(), "💡 Want to create a group chat? You totally can! Bring your friends together and vibe out in your own space! 🔥👥");
-        greetingMessage(newUser.getUserId(), "🛍️ Oh, and did I mention the Marketplace? You can list items for sale—privately for friends or publicly for everyone to see! Buyers can message you directly, and you can search for cool deals too! 💰🛒");
-        greetingMessage(newUser.getUserId(), "🌙 Prefer a dark mode? Check out the settings and switch up the vibe! 😎🌌");
-        greetingMessage(newUser.getUserId(), "That's it for now! 🎯 Start exploring, make connections, and enjoy your time on VIBEZ! 🚀💜");
         return true;
-    }
-
-    public void greetingMessage(String userId, String messageText) {
-        Message message = new Message();
-        message.setSenderId("679b99e09b263c53d878d30f");
-        message.setReceiverId(userId);
-        message.setMessage(messageText);
-        message.setTimestamp(java.time.LocalDateTime.now());
     }
 
     public String authenticateUser(User existingUser) {
